agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-google>=0.1.0
python-dotenv
livekit-plugins-difyai
livekit-plugins-deepgram